Add XML image analysis report and batch export
- Introduced XMLDump module for detailed XML reports of disk images, including metadata, hashes, summary, and directory structure.
- Added "Image Analysis (XML)" and "Batch Image Analysis (XML)" menu items with localization.
- Implemented handlers for generating and displaying/saving XML reports for single or multiple images.
- Replaced Win9xCleanScanner with LoadedImageScanner for batch operations.
- Enhanced ItemScanForm to return cancellation/error status and show decimal progress.
- Updated IImageScanner and FilterScanner to use Double for progress.
- Added IsFileEmpty to DirectoryEntryBase for improved empty entry detection.
- Made relevant SummaryPanel methods Shared for reuse.
- Updated project, resources, and designer files for new features and UI tweaks.
- Bumped version to 2.37.0.

+++ DiskImageTool/Scanners/IImageScanner.vb
@@ -1,7 +1,7 @@
 ﻿Imports System.Threading
 
 Public Interface IImageScanner
-    Event ProgressChanged(Percent As Integer)
+    Event ProgressChanged(Percent As Double)
     Event ScanCompleted(Cancelled As Boolean, [Error] As Exception)
 
     ReadOnly Property ItemsRemaining As UInteger
